This commit fixes a bug that would prevent the alert from autokilling properly.
The bug came from the potential differences between the set configuration and the custom config given for the alert. The function that manages the removal can take an alert object as a parameter, and that object contains the custom config. The bug was that that removal function was not given the alert, so it fell back to the set config, and the target set in that was not the parent of the alert.

Also, some documentation improvements and switched logging off.